# Conflicts:
#	src/main/java/com/rivalhub/common/Config.java
#	src/main/java/com/rivalhub/common/ErrorMessages.java

+++ src/main/java/com/rivalhub/organization/OrganizationStationService.java
@@ -4,39 +4,45 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.fge.jsonpatch.JsonPatchException;
 import com.github.fge.jsonpatch.mergepatch.JsonMergePatch;
 import com.rivalhub.common.AutoMapper;
+import com.rivalhub.common.FormatterHelper;
 import com.rivalhub.common.MergePatcher;
 import com.rivalhub.event.EventType;
 import com.rivalhub.organization.exception.OrganizationNotFoundException;
-import com.rivalhub.reservation.AddReservationDTO;
-import com.rivalhub.reservation.ReservationValidator;
-import com.rivalhub.station.*;
+import com.rivalhub.station.EventTypeStationsDto;
+import com.rivalhub.station.Station;
+import com.rivalhub.station.StationAvailabilityFinder;
+import com.rivalhub.station.StationDTO;
 import com.rivalhub.user.UserData;
-import com.rivalhub.user.UserNotFoundException;
-import com.rivalhub.user.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class OrganizationStationService {
+
     private final RepositoryManager repositoryManager;
     private final AutoMapper autoMapper;
     private final MergePatcher<StationDTO> stationMergePatcher;
     private final OrganizationStationValidator validator;
 
     StationDTO addStation(StationDTO stationDTO, Long id, String email) {
-        Organization organization = repositoryManager.findOrganization(id);
-
-        UserData user = repositoryManager.findUser(email);
+        UserData user = repositoryManager.findUserByEmail(email);
         List<Organization> organizationList = user.getOrganizationList();
-        organizationList
+
+        Organization organization = organizationList
                 .stream().filter(org -> org.getId().equals(id))
                 .findFirst()
                 .orElseThrow(OrganizationNotFoundException::new);
+
+        OrganizationSettingsValidator.checkIfUserIsAdmin(user, organization);
 
         Station station = autoMapper.mapToStation(stationDTO);
         UserOrganizationService.addStation(station, organization);
@@ -49,7 +55,7 @@ public class OrganizationStationService {
 
     List<Station> viewStations(Long organizationId, String start, String end, EventType type,
                                boolean onlyAvailable, String email, boolean showInactive) {
-        UserData user = repositoryManager.findUser(email);
+        UserData user = repositoryManager.findUserByEmail(email);
 
         Organization organization = validator.checkIfViewStationIsPossible(organizationId, user);
 
@@ -57,66 +63,77 @@ public class OrganizationStationService {
         if (!showInactive) stationList = filterForActiveStations(stationList);
 
         if (onlyAvailable && start != null && end != null)
-            return getAvailableStations(organization, start, end, type, user, stationList);
+            return StationAvailabilityFinder
+                    .getAvailableStations(organization, start, end, type, user, stationList);
 
         return stationList;
     }
 
-    public List<Station> getAvailableStations(Organization organization, String startTime,
-                                              String endTime, EventType type, UserData user, List<Station> stationList) {
-        List<Station> availableStations = new ArrayList<>();
+    List<EventTypeStationsDto> getEventStations(Long organizationId, String start, String end, EventType type) {
+        UserData userData = repositoryManager
+                .findUserByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
 
-        stationList.forEach(station -> {
-            AddReservationDTO reservationDTO = new AddReservationDTO();
-            reservationDTO.setStartTime(startTime);
-            reservationDTO.setEndTime(endTime);
-            reservationDTO.setStationsIdList(List.of(station.getId()));
+        Organization organization = repositoryManager.findOrganizationById(organizationId);
+        List<Station> availableStations = StationAvailabilityFinder
+                .getAvailableStations(organization, start, end, type, userData, organization.getStationList());
+        availableStations = filterForActiveStations(availableStations);
 
-            if (type != null && !station.getType().equals(type)) {
-                return;
-            }
+        LocalDateTime startTime = LocalDateTime.parse(start, FormatterHelper.formatter());
+        LocalDateTime endTime = LocalDateTime.parse(end, FormatterHelper.formatter());
 
-            if (ReservationValidator.checkIfReservationIsPossible(
-                    reservationDTO,
-                    organization,
-                    user,
-                    organization.getId(),
-                    List.of(station))) {
+        List<EventTypeStationsDto> eventStations = new ArrayList<>();
+        Duration timeNeeded = Duration.ofSeconds(ChronoUnit.SECONDS.between(startTime, endTime));
 
-                availableStations.add(station);
-            }
-        });
+        if (type != null) {
+            eventStations.add(getEventTypeStation(type, availableStations, organization, timeNeeded));
+            return eventStations;
+        }
 
-        return availableStations;
+        for (EventType eventType : EventType.values()) {
+            eventStations.add(getEventTypeStation(eventType, availableStations, organization, timeNeeded));
+        }
+
+        return eventStations;
     }
 
-    StationDTO findStation(Long stationId){
+    private EventTypeStationsDto getEventTypeStation(EventType eventType,
+                                                     List<Station> availableStations,
+                                                     Organization organization,
+                                                     Duration timeNeeded) {
+
+        EventTypeStationsDto eventStation = new EventTypeStationsDto();
+        eventStation.setType(eventType);
+        eventStation.setStations(availableStations.stream().map(autoMapper::mapToNewStationDto).toList());
+        eventStation.setFirstAvailable(StationAvailabilityFinder
+                .getFirstDateAvailableForDuration(organization.getStationList(), timeNeeded));
+
+        return eventStation;
+    }
+
+    StationDTO findStation(Long stationId) {
         return autoMapper.mapToNewStationDto(repositoryManager.findStationById(stationId));
     }
 
     public void updateStation(Long organizationId, Long stationId, String email, JsonMergePatch patch) throws JsonPatchException, JsonProcessingException {
-        UserData user = repositoryManager.findUser(email);
+        UserData user = repositoryManager.findUserByEmail(email);
 
         validator.checkIfUpdateStationIsPossible(organizationId, user);
 
         StationDTO station = findStation(stationId);
         StationDTO stationPatched = stationMergePatcher.patch(patch, station, StationDTO.class);
         stationPatched.setId(stationId);
-        updateStation(stationPatched);
-    }
 
-    void updateStation(StationDTO stationDTO){
-        Station station = autoMapper.mapToStation(stationDTO);
-        repositoryManager.save(station);
+        repositoryManager.save(autoMapper.mapToStation(stationPatched));
     }
 
     @Transactional
-    public void deleteStation(Long stationId, Long organizationId) {
+    public void deleteStation(Long stationId, Long organizationId, String email) {
         Organization organization = repositoryManager.findOrganizationById(organizationId);
+        OrganizationSettingsValidator.checkIfUserIsAdmin(repositoryManager.findUserByEmail(email), organization);
         UserOrganizationService.removeStation(repositoryManager.findStationById(stationId), organization);
     }
 
-    private List<Station> filterForActiveStations(List<Station> stationList){
+    private List<Station> filterForActiveStations(List<Station> stationList) {
         return stationList
                 .stream().filter(Station::isActive)
                 .toList();
